Add option to automatically determine the overscan axis
If the overscan axis is explicitly set to `None` in
`subtract_overscan`, the overscan direction is determined from the
shape of the overscan section, set the shortest dimension. (In case of
a square overscan, the default of 1 is picked.)

This can be convenient in cases where `fits_section` is used with an
(unknown) input string (for example, from an image's BIASSEC header
keyword): the user can let the function determine the appropriate
axis.