Add a 'sort' method to ImageFileCollection
This uses the underlying table structure to do that sort, it also updates
the separate self._files accordingly.
This method allows me to set a the order in which the files are processed,
e.g. I can iterate over all files and make plots sorted alphabetically by
'OBJECT' or by 'OBS-TIME' or 'FILTER'.

A related chage is to sort the file list when reading files from a
directory. This sorting is alphapetical, if the user does not like it
they can always use the new 'sort' function later.
This change makes the order of files that are read in reproducable, which useful
when the iterators are used to make plots of each file - I want the same order
when I re-run my script.
(Without this change the order of files seems to change randomly at times.)