Make meta and header attributes separate objects
The meta and header attributes could be interchangeably used, either
as a fits.Header or a dict. This commit changes that, so that the meta
attribute is always a dict, and the header attribute is always a
fits.Header. Keywords in meta are *always* lower-case.

Under the hood, one is automatically converted to the other one (as
far as possible without loss of information) using properties. Every
time one of the two is accessed, a _dirty attribute is set to mark
_meta or _header as possible changed (accessing doesn't mean it has
changed, but it may have).

Comments and history cards get stored as lists in the meta attribute,
and vice versa converted to relevant fits.Header cards (both 'comment'
and 'comments' keys are accepted. Note that all keys in meta,
'history', 'comment' or 'comments', should be lower-case).

Depending on whether the CCDData object was initialized with a
fits.Header object or a dict object, the original object can always be
retrieved using the relevant attribute.