ImageFileCollection fails if the headers contain the keyword 'FILE'
There's an assert statement on line 503:
assert 'file' not in h
to catch and raise the problem - but no handling is implemented. 
If not caught the 'file' value list in the summary dictionary would get a filename appended twice for each image.
This will cause the generation of the summary_table to fail.

The suggested way to fix this is to remove the header 'FILE' keyword if present.
This should be okay since it will either be the same as the input filename or likely incorrect - I guess?