Using temporary files in tests [#736]
This commit addresses issue [#736], where `test_image_collection.py`
in the test suite generates a `test_table.txt` in the working directory
and does not clean up after itself. This is fixed now by using a
`NamedTemporaryFile` within a context manager, and deletes the file
once the test is completed.