Fix bug: ImageFileCollection.filter() would return an empty collection
If the filtering criteria passed to ImageFileCollection.filter() would return
an empty collection, a crash occurs because the method instantiates a new
ImageFileCollection() with a list of filenames returned from .files_filtered().
When a new instance of ImageFileCollection() is called without a filenames
list, _get_files() -> _fits_files_in_directory() attempts to create a list
of all FITS files within the directory contained in self.location.  If this
new instance of ImageFileCollection() is called without setting location, then
self.location is set to "".  When _fits_files_in_directory() attempts to do a
directory listing via listdir(), an error is raised:
    FileNotFoundError: [Errno 2] No such file or directory: ''

This commit adds a check for empty self.location before any attempted call to
_fits_files_in_directory().  The method _get_files() simply returns an empty
string, leading to an empty ImageFileCollection instance.  This is the expected
behavior for filtering a collection such that no files meet the criteria.

	modified:   ccdproc/image_collection.py