Escape parentheses in FITS keyword values used for IFC filtering
In `ImageFileCollection()`, the method `_find_keywords_by_values()` uses regular
expressions to find matches for specified FITS keywords by value.  In the case
where a FITS keyword value has parentheses (whether for a mathematical or
facility reason), the regex search interprets the parenthesis as a control
character and does not correctly match the string as specified, returning none
of the desired frames.

As an example from the Lowell Observatory Discovery Telescope (LDT) DeVeny
Optical Spectrograph, we specify order-blocking filters both by name and filter
wheel position, e.g.:
FILTREAR= 'Clear (C)'          / Rear Filter

There is logic in the `_find_keywords_by_values()` method to specify that one is
supplying a regular expression for search, but the method also constructs a
simple regex search based on the specified keyword value if `regex_match=False`.

This commit adds escape characters to parentheses in the case where the user
has NOT specified an input regular expression.  The resulting pattern search is
successful.

	modified:   ccdproc/image_collection.py