Change manual escape of parentheses to re.escape()
The original PR for this issue [#770] manually added escapes for parentheses.
As it turns out, python is smarter than I am and already has a function to find
and escape all regex special characters in a string.

This change will allow any regex special character to exist in keyword filtering
requests to `ImageFileCollection()`.

	modified:   ccdproc/image_collection.py