Add test for escaping regex special characters for filtering
In ``ImageFileCollection()``, filtering by keyword is done with use of regular
expressions.  In PR #770, added escaping of any regex special characters that
might be present in the keyword used for filtering a collection.  This test
checks that keywords containing regex special characters are correctly escaped
and therefore properly return the desired filtered collection.

	modified:   CHANGES.rst
	modified:   ccdproc/tests/test_image_collection.py