Add `HDUList` generator to `ImageFileCollection`
Presently, the `ImageFileCollection` generator `hdus` states in the
documentation that it returns the `HDUList` for each file in the collection,
but in fact it returns only the HDU for the extension specified in the
initialization of the collection.

This commit clarifies that the `hdus` generator returns either a ImageHDU
or BinTableHDU for the specified extension.  Additionally, it adds a new
`hduls` generator that actually returns the HDUList for the entire file.

	modified:   ccdproc/image_collection.py
	modified:   docs/image_management.rst