Change `.hduls()` to `.hdulists()` for clarity; conformal changes
For clarity, make the new generator be called `.hdulists()` to make it more
distincty from `.hdus()`, and be very clear that it returns HDUList objects
and not just HDUs.

Remove `overwrite=True` option from `.hdulists()` generator

Within the current framework of `ImageFileCollection._generator()`, it would be
ery challenging to allow for `overwrite=True` for the new HDUList generator.

Because the `overwrite=True` option is "living dangerously" to begin with,
simply opt to not provide the functionality.

Add unit test to ensure we cannot overwrite from the new `.hdulists()`
generator.

Fix typo in doc, and include PR# where needed

	modified:   CHANGES.rst
	modified:   ccdproc/image_collection.py
	modified:   ccdproc/tests/test_image_collection.py
	modified:   docs/image_management.rst